# docs/sacp/openapi-agent.yaml
openapi: 3.0.3
info:
  title: WSIM Agent Commerce API
  description: |
    WSIM Agent Commerce API - Enables AI agents to make authenticated purchases on behalf of users.

    ## Overview

    This API supports the SimToolBox Agent Commerce Protocol (SACP), allowing AI agents to:
    - Register with user wallets via OAuth Device Authorization or pairing codes
    - Authenticate using OAuth 2.0 client credentials
    - Request payment tokens for purchases
    - Handle step-up authorization for transactions exceeding limits

    ## Authorization Flows

    ### Flow 1: Device Authorization Grant (RFC 8628) - RECOMMENDED
    Agent-initiated flow for autonomous agents:
    1. Agent calls `POST /api/agent/v1/oauth/device_authorization`
    2. Receives `device_code` and `user_code` (format: WSIM-XXXXXX-XXXXXX)
    3. Agent displays `user_code` and `verification_uri` to user
    4. User enters code in WSIM mobile app
    5. Agent polls `POST /api/agent/v1/oauth/token` with `grant_type=urn:ietf:params:oauth:grant-type:device_code`
    6. On approval, receives `client_id` and `client_secret`
    7. Uses client credentials to obtain access tokens

    ### Flow 2: Pairing Code (User-Initiated)
    User-initiated flow when user has pre-generated code:
    1. User generates pairing code in mobile app
    2. User provides code to agent (WSIM-XXXXXX-XXXXXX format)
    3. Agent calls `POST /api/agent/v1/access-request` with the code
    4. Agent polls until user approves
...

...

...

...

...

...

...

...
  contact:
...
servers:
...
paths:
...
  /api/agent/v1/payments/token:
    post:
      tags:
        - Payments
      summary: Request payment token
      description: |
        Request a payment token for a purchase. The token can be used with the
        payment network (NSIM) to complete the transaction.

        **Spending Limit Check**:
        - If amount is within per-transaction limit: Token issued immediately
        - If amount exceeds limit: Step-up authorization required

        **Step-Up Flow**:
        When step-up is required, the response includes `step_up_required: true`
        and a `poll_url`. The agent should:
        1. Present step-up information to the user (or wait for push notification)
        2. Poll the `poll_url` (e.g., `GET /api/agent/v1/payments/{payment_id}/status`) every 2-5 seconds
        3. Once approved, the status response includes the `payment_token`

        **Payment Method Selection**:
        - If `payment_method_id` is omitted, uses user's default payment method
        - For step-up approvals, user can select a different payment method
      operationId: requestPaymentToken
      security:
        - agentBearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PaymentTokenRequest'
      responses:
        '200':
          description: Payment token issued or step-up required
          content:
            application/json:
              schema:
                oneOf:
                  - $ref: '#/components/schemas/PaymentTokenResponse'
                  - $ref: '#/components/schemas/StepUpRequiredResponse'
        '400':
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Invalid or expired agent token
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Agent suspended or over spending limit
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '429':
          description: Rate limit exceeded
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/agent/v1/payments/{paymentId}/status:
    get:
      tags:
        - Payments
      summary: Check payment status
      description: |
        Check the status of a payment request, particularly useful for:
        - Polling during step-up authorization
        - Verifying payment completion

        **Polling Guidance**:
        - Poll every 2-5 seconds during step-up wait
        - Stop polling after 15 minutes (step-up expiration)
        - Use webhook subscription for real-time updates (Phase 2)
      operationId: getPaymentStatus
      security:
        - agentBearerAuth: []
      parameters:
        - name: paymentId
          in: path
          required: true
          schema:
            type: string
          description: Payment request ID
      responses:
        '200':
          description: Payment status
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaymentStatusResponse'
        '404':
          description: Payment not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/agent/v1/payments/methods:
    get:
      tags:
        - Payments
      summary: List available payment methods
      description: |
        Returns all payment methods available to the agent for purchases.
        This includes all cards enrolled in the owner's wallet.

        The `is_default` flag indicates the user's default payment method
        which is used for auto-approved transactions.
      operationId: listPaymentMethods
      security:
        - agentBearerAuth: []
      responses:
        '200':
          description: List of payment methods
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaymentMethodsResponse'
        '401':
          description: Invalid or expired agent token
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  # =============================================================================
  # Agent Management Endpoints (Mobile)
  # =============================================================================
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Mobile app JWT authentication

    agentBearerAuth:
      type: http
      scheme: bearer
      description: Agent OAuth access token

    basicAuth:
      type: http
      scheme: basic
      description: Basic auth for introspection (client_id:client_secret)
  schemas:
...
    PaymentTokenRequest:
      type: object
      required:
        - merchant_id
        - amount
        - currency
      properties:
        merchant_id:
          type: string
          description: Merchant identifier (SSIM store ID)
          example: "store_xyz789"
        session_id:
          type: string
          description: Checkout session ID from merchant
          example: "sess_abc123"
        amount:
          type: number
          format: decimal
          description: Transaction amount
          example: 156.48
        currency:
          type: string
          pattern: "^[A-Z]{3}$"
          description: ISO 4217 currency code
          example: "CAD"
        payment_method_id:
          type: string
          description: Specific payment method to use (optional, defaults to user's default)
          example: "pm_card_visa_1234"
        items:
          type: array
          description: Cart items (for step-up display)
          items:
            $ref: '#/components/schemas/CartItem'
        mandate_type:
          type: string
          enum: [cart, intent, none]
          description: Type of mandate (Phase 2)
          default: cart
        callback_url:
          type: string
          format: uri
          description: URL for step-up result callback (optional)
    CartItem:
      type: object
      properties:
        name:
          type: string
          example: "Canadian Maple Syrup"
        quantity:
          type: integer
          example: 2
        price:
          type: number
          format: decimal
          example: 24.99
    PaymentTokenResponse:
      type: object
      description: Response when payment token is issued (no step-up required)
      required:
        - step_up_required
        - payment_id
        - payment_token
        - expires_in
        - payment_method
        - next_step
      properties:
        step_up_required:
          type: boolean
          enum: [false]
          description: Indicates no step-up was required
        step_up_id:
          type: string
          nullable: true
          description: Always null when step_up_required is false
          example: null
        payment_id:
          type: string
          description: Payment/transaction ID
          example: "550e8400-e29b-41d4-a716-446655440000"
        payment_token:
          type: string
          description: JWT token to use with merchant for payment
          example: "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9..."
        expires_in:
          type: integer
          description: Token lifetime in seconds
          example: 300
        payment_method:
          type: object
          description: Payment method used
          required:
            - id
            - type
            - last_four
          properties:
            id:
              type: string
              example: "pm_card_visa_1234"
            type:
              type: string
              example: "Visa"
            last_four:
              type: string
              example: "4242"
        next_step:
          type: string
          description: Human-readable guidance for the AI agent on what to do next
          example: "Use payment_token with merchant to complete purchase"
    StepUpRequiredResponse:
      type: object
      required:
        - payment_id
        - status
        - step_up_required
        - step_up_id
        - step_up_expires_at
        - reason
        - poll_url
        - next_step
      properties:
        payment_id:
          type: string
          description: Payment request ID (use this in the poll_url)
          example: "pay_abc123xyz789"
        status:
          type: string
          enum: [awaiting_authorization]
          description: Payment status
        step_up_required:
          type: boolean
          enum: [true]
          description: Indicates step-up is required
        step_up_id:
          type: string
          description: Step-up request ID for tracking
          example: "stepup_xyz789abc123"
        step_up_expires_at:
          type: string
          format: date-time
          description: Step-up expiration (15 minutes)
          example: "2026-01-21T15:15:00Z"
        reason:
          type: string
          description: Human-readable reason for step-up (only present when status is pending/awaiting_authorization)
          example: "Amount exceeds per-transaction limit of $100.00 CAD"
        trigger_type:
          type: string
          enum: [per_transaction, daily_limit, monthly_limit]
          description: What limit triggered the step-up (only present when status is pending/awaiting_authorization)
        poll_url:
          type: string
          format: uri
          description: |
            URL to poll for payment status. Poll every 2-5 seconds until
            status changes from 'awaiting_authorization' to 'approved' or 'rejected'.
          example: "https://wsim.banksim.ca/api/agent/v1/payments/pay_abc123xyz789/status"
        next_step:
          type: string
          description: Human-readable guidance for the AI agent on what to do next
          example: "Poll poll_url until status changes to approved or rejected"
...
    PaymentMethodsResponse:
      type: object
      required:
        - payment_methods
      properties:
        payment_methods:
          type: array
          items:
            $ref: '#/components/schemas/PaymentMethod'
    PaymentMethod:
      type: object
      description: Payment method available for agent purchases
      required:
        - id
        - type
        - last_four
        - is_default
      properties:
        id:
          type: string
          description: Payment method ID
          example: "pm_card_visa_1234"
        type:
          type: string
          description: Card type/brand (e.g., "Visa", "Mastercard")
          example: "Visa"
        last_four:
          type: string
          description: Last 4 digits of card
          example: "4242"
        cardholder_name:
          type: string
          description: Name on the card
          example: "JOHN DOE"
        expiry_month:
          type: integer
          description: Expiration month (1-12)
          example: 12
        expiry_year:
          type: integer
          description: Expiration year (4-digit)
          example: 2027
        is_default:
          type: boolean
          description: Whether this is the user's default payment method
          example: true

    # =========================================================================
    # Agent Management Schemas
    # =========================================================================
...
    ErrorResponse:
      type: object
      required:
        - error
      properties:
        error:
          type: object
          required:
            - code
            - message
          properties:
            code:
              type: string
              description: Machine-readable error code
              example: "AGENT_NOT_FOUND"
            message:
              type: string
              description: Human-readable error message
              example: "Agent not found"
            details:
              type: object
              description: Additional error context
...